requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
craft-text-detector==0.4.3
cycler==0.11.0
filelock==3.12.2
fonttools==4.42.0
fpdf==1.7.2
gdown==4.7.1
idna==3.4
imageio==2.31.1
imgproc==0.0.2.1
importlib-resources==6.0.1
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
poppler-utils==0.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sympy==1.12
tifffile==2023.7.10
torch==1.10.0
torchvision==0.8.2
tqdm==4.66.0
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
python-dotenv
transformers
accelerate
